ShockDetection.jl

ShockDetection.jl is a Julia package for detecting shocks in time series data from Godunov solver.

Features

A method for shock detection.

Easy-to-use functions for analyzing time series data.

Compatible with Julia 1.9.

Installation

You can install ShockDetection.jl using the Julia package manager. From the Julia REPL, type ] to enter the Pkg REPL mode

Functions

ShockwaveDetection.read_output_file — Function

read_output_file(filename)

Reads data from an output file.

Arguments

filename::String: The path to the output file.

Returns

x0_xmax::Vector{Float64}: A vector containing the first and last x values.

t_values::Vector{Float64}: A vector containing the time values.

u_values::Array{Float64, 3}: A 3D array containing the u values.

dims_u::Vector{Int}: A vector containing the dimensions of the u values (Nu x Nx).

This function reads data from the specified output file generated from 1d_plots.jl that use Euler2D.jl to solve, which is assumed to have a specific format. It reads the dimensions of the u values, the first and last x values, the number of time steps, the time values, and the u values themselves. It reshapes the u values into a 3D array and returns all the read data.

ShockwaveDetection.convert_to_primitive — Function

convert_to_primitive(u_values)

Converts conserved state variables to primitive variables.

Arguments

u_values::Array{Float64, 3}: A 3D array containing the conserved state variables (density, momentum, total energy) at each point x and time t.

Returns

density_field::Array{Float64, 2}: A 2D array containing the density field.

velocity_field::Array{Float64, 2}: A 2D array containing the velocity field.

pressure_field::Array{Float64, 2}: A 2D array containing the pressure field.

Euler equations are typically represented in a "conserved" form, where the vector u contains (density, momentum, total energy) at each point x and time t. This function converts these conserved state variables to primitive variables (density, velocity, pressure) in order to calculate delta1 and delta2 according to [6].

The u_values parameter is a 3D array representing the conserved state variables at each point x and time t. This function iterates over each time step and calculates the primitive state vector for density, velocity, and pressure using appropriate transformations. The resulting primitive variables are stored in separate arrays density_field, velocity_field, and pressure_field, and returned.

ShockwaveDetection.detect_shock — Function

detect_shock(density_field, velocity_field, pressure_field, x0_xmax, t_values; threshold=0.5)

Detects shock positions over time based on given density, velocity, and pressure fields.

Arguments

density_field::Matrix: Matrix representing the density field over space and time.

velocity_field::Matrix: Matrix representing the velocity field over space and time.

pressure_field::Matrix: Matrix representing the pressure field over space and time.

x0_xmax::Tuple: Tuple containing the start and end positions of the spatial domain.

t_values::Vector: Vector containing time values.

threshold::Float64: Threshold value for detecting shocks. Default is 0.5.

Returns

shock_positions_over_time::Vector: Vector containing shock positions at each time step.

Details

This function iterates over each time step and detects shock positions using the detect_shocks_at_timestep function.

ShockwaveDetection.create_wave_animation — Function

create_wave_animation(x0_xmax, t_values, density_field, velocity_field, pressure_field)

Create an animation of density, velocity, and pressure fields over time.

Arguments

x0_xmax::Tuple{Float64, Float64}: A tuple containing the minimum and maximum x values.

t_values::Array{Float64}: An array of time values.

density_field::Array{Float64, 2}: A 2D array representing the density field. Each column represents the density at a different time.

velocity_field::Array{Float64, 2}: A 2D array representing the velocity field. Each column represents the velocity at a different time.

pressure_field::Array{Float64, 2}: A 2D array representing the pressure field. Each column represents the pressure at a different time.

Returns

This function saves an animation as a gif file named "densityvelocitypressureovertime.gif" in the current directory and returns nothing.

ShockwaveDetection.create_wave_animation_with_shock — Function

create_wave_animation_with_shock(x, t_values, density_field, velocity_field, pressure_field, shock_positions_over_time; save_file=false)

Create an animation of density, velocity, and pressure fields over time, with markers for shock positions.

Arguments

x0_xmax::Tuple{Float64, Float64}: A tuple containing the minimum and maximum x values.

t_values::Vector: Time values.

density_field::Matrix: Density field over space and time.

velocity_field::Matrix: Velocity field over space and time.

pressure_field::Matrix: Pressure field over space and time.

shock_positions_over_time::Vector{Vector{Int}}: Shock positions over time.

save_file::Bool: Whether to save the animation as a gif file (default: false).

Returns

anim::Animation: Animation object.

This function creates an animation of density, velocity, and pressure fields over time, with markers indicating shock positions at each time step. The animation is generated using the Plots.jl package and returned as an Animation object. If save_file is true, the animation is saved as a gif file with the filename 'densityvelocitypressureovertimewithshock_positions.gif'.