ShockDetection.jl

ShockDetection.jl is a Julia package for detecting shocks in time series data from Godunov solver. The packages can detect shock discontinuities for grid data with dimensions

Nx x Nt when Nx are the number of grid points in x axis and Nt is the number of time steps in the simulation

Nx x Ny x Nt when Nx and Ny are the number of grid points in x and y axes respectively, and Nt is the number of time steps in the simulation

The package can only process simulation files that end with .tape and .celltape. Other simulation file types are not supported!

Features

A method for shock detection. The gradient of each point in the simulation is calculated using finite difference method, clustered by DBSCAN algorithm and optimal shock curves are fitted using Levenberg–Marquardt algorithm.

Visualization is also available within the package. You can also automatically calculate the normal vectors of 2D simulation as you call the function create_heatmap_evo_with_shock. But if you wish to obtain the normal vector of each curve in a specified time frame, the function calculate_normal_vector is also available for you.

Compatible with Julia 1.9 and 1.10.

Installation

You can install ShockDetection.jl using the Julia package manager. From the Julia REPL, type ] to enter the Pkg REPL mode There are several dependencies needed to run this pacakge. Checkout Project.toml to see what packages you need to have in order to run ShockwaveDetection.jl

Guide

Go to our examples directory to see how to use our functions!

Short guide on how to use our package on grid data with dimension Nx x Ny x Nt

When you wish to use our package to detect shock discontinuities, you should first input your data file path into FlowData. After that, define your algorithm of finite difference, clustering, and fitting that best suited your simulation data and use the function detect to automate the process the detection process. The results of your detection can be found in ShockDetectionResult2D. You can further use various functions in visualization to see the result.

Functions and structs

Main functions and structs of our package

These are the most important functions and structs in our package. Others are functions/structs called/used/stored by detect/ ShockDetectionResult2D to perform shock detections but they still exist in this documentation to clarify to you how our algorithm works!

ShockwaveDetection.detect — Function

detect(flow_data::FlowData, shock_point_algo::Abstract2DShockDetectionAlgo, cluster_algo::DBSCANAlgo)

Detects shocks in 2D flow data, clusters the shockpoints and applies fitting to the cluster. It also shows the runtime and memory allocations required in each subprocess

Arguments

flow_data::FlowData: A FlowData object containing the 2D flow field data.

shock_point_algo::Abstract2DShockDetectionAlgo: An algorithm for detecting shock points in 2D flow data.

cluster_algo::DBSCANAlgo: A clustering algorithm (e.g., DBSCAN) to group detected shock points into clusters.

Returns

ShockDetectionResult2D: An object containing:

shock_positions_over_time: Detected shock points over time.

shock_clusters_over_time: Clusters of shock points over time.

shock_fits_over_time: Fitted shock curves over time.

Description

This function detects shock points in 2D flow data using a specified shock detection algorithm. Detected shock points are clustered using the provided DBSCANAlgo, and then the clusters are fitted to create a smooth representation of the shock over time.

detect(flow_data::FlowData, shock_point_algo::Abstract1DShockDetectionAlgo)

Detects shocks in 1D flow data and show the runtime and memory allocations required in each subprocess.

Arguments

flow_data::FlowData: A FlowData object containing the 1D flow field data.

shock_point_algo::Abstract1DShockDetectionAlgo: An algorithm for detecting shock points in 1D flow data.

Returns

ShockDetectionResult1D: An object containing:

shock_positions_over_time: Detected shock points over time.

Description

This function detects shock points in 1D flow data using a specified shock detection algorithm.

ShockwaveDetection.ShockDetectionResult2D — Type

struct ShockDetectionResult2D <: AbstractShockDetectionResult

A structure to hold the results of 2D shock detection.

Fields

shock_positions_over_time::Vector{Any}: A vector of the CartesianIndices of shock points detected in every time frame.

shock_clusters_over_time::Vector{Any}: A vector of groups of clustered shock points detected in every time frame.

shock_fits_over_time::Vector{Any}: Fitted shock curves over time. In each time frame is a vector of objects Fitting

Description

This structure is used to store the results of shock detection in 2D flow data. It contains the detected shock points, clusters of shock points, and fitted shock curves over time.

Input-related structs and functions

ShockwaveDetection.convert_to_primitive — Function

convert_to_primitive(u_values)

Converts conserved state variables to primitive variables.

Arguments

u_values::Array{Float64, 3}: A 3D array containing the conserved state variables (density, momentum, total energy) at each point x and time t.

Returns

density_field::Array{Float64, 2}: A 2D array containing the density field.

velocity_field::Array{Float64, 2}: A 2D array containing the velocity field.

pressure_field::Array{Float64, 2}: A 2D array containing the pressure field.

Euler equations are typically represented in a "conserved" form, where the vector u contains (density, momentum, total energy) at each point x and time t. This function converts these conserved state variables to primitive variables (density, velocity, pressure).

The u_values parameter is a 3D array representing the conserved state variables at each point x and time t. This function iterates over each time step and calculates the primitive state vector for density, velocity, and pressure using appropriate transformations. The resulting primitive variables are stored in separate arrays density_field, velocity_field, and pressure_field, and returned.

ShockwaveDetection.FlowData — Type

struct FlowData{N, T}

A structure to hold flow data for simulations.

Fields

ncells::NTuple{N, Int}: Number of cells in each dimension.

nsteps::Int: Number of time steps in the simulation.

bounds::NTuple{N, Tuple{T, T}}: Bounds of the simulation domain in each dimension.

tsteps::Vector{T}: Time steps of the simulation.

u::Union{Array{T}, Nothing}: Array of primitive variables or Nothing for .celltape data.

density_field::Array{T}: Density field of the simulation.

velocity_field::Array{T}: Velocity field of the simulation.

pressure_field::Array{T}: Pressure field of the simulation.

mach_to_m_s::Bool: Flag indicating if Mach number should be converted to meters per second.

cell_ids::Union{Matrix{Int64}, Nothing}: Cell IDs for .celltape data. Nothing for .tape data.

Finite difference-related functions and structs

ShockwaveDetection.ImageProcessingShockDetectionAlgo — Type

struct ImageProcessingShockDetectionAlgo <: Abstract2DShockDetectionAlgo

A structure that represents the image processing shock detection algorithm.

Fields

threshold::Float64: The threshold value for detecting shocks.

kernelname::Symbol: The name of the kernel to be used for image processing (e.g., :sobel, :prewitt).

Description

This struct is used to configure the image processing shock detection algorithm. The algorithm uses a specified kernel to process the image and detect shocks based on the provided threshold value.

Example

algo = ImageProcessingShockDetectionAlgo(threshold=0.5, kernelname=:sobel)

Cluster-related functions and structs

ShockwaveDetection.cartesian_index_to_xy — Function

cartesian_index_to_xy(shock_positions_t, x, y) -> Matrix

Converts Cartesian indices of shock positions into their corresponding x and y coordinates.

Arguments

shock_positions_t: An array of CartesianIndex indicating the positions of shocks.

x: An array representing the x-coordinates in the domain.

y: An array representing the y-coordinates in the domain.

Returns

Matrix: A 2xN matrix where the first row contains the x-coordinates and the second row contains the y-coordinates of the shock positions. The type of the elements in the matrix matches the type of elements in x and y.

ShockwaveDetection.cluster_shock_points — Function

cluster_shock_points(dbscan_algo::DBSCANAlgo, shock_positions_over_time, flow_data)

A part of the 2D detection algorithm of function detect. This function clusters shock points over time using the DBSCAN algorithm.

Arguments

dbscan_algo::DBSCANAlgo: An instance of the DBSCAN clustering algorithm.

shock_positions_over_time: A vector containing shock positions for each time step.

flow_data: A FlowData object containing the flow field data.

Returns

shock_clusters_over_time: A vector containing clustered shock points for each time step.

Description

This function clusters shock points over time using the DBSCAN algorithm. It iterates over each time step, converts shock positions to Cartesian coordinates, and applies the DBSCAN algorithm to cluster the shock points. The clustered shock points are stored in a vector, with each element corresponding to a time step.

ShockwaveDetection.DBSCANAlgo — Type

struct DBSCANAlgo

A structure that keeps all the necessary parameters for the DBSCAN clustering algorithm.

Fields

radius::Float64: The radius within which to search for neighboring points.

min_neighbors::Int: The minimum number of neighbors required to form a dense region.

min_cluster_size::Int: The minimum number of points required to form a cluster.

Constructors

DBSCANAlgo(radius::Float64 = 0.5, min_neighbors::Int = 3, min_cluster_size::Int = 10): Creates a new DBSCANAlgo instance with the specified parameters. Default values are provided for all parameters.

Example

algo = DBSCANAlgo(radius=1.0, minneighbors=5, mincluster_size=15)

Fitting-related functions and structs

ShockwaveDetection.fit_shock_clusters_over_time — Function

fit_shock_clusters_over_time(shock_clusters_over_time)

A part of the 2D pipeline of detect. In each time frame which is stored in the array shock_clusters_over_time, this function fits the shock clusters to existing models of parametized curves using the fit_shock_clusters function. The fitted shock clusters are stored in the array shock_fits_over_time.

Arguments

shock_clusters_over_time: An array of shock clusters over time.

Returns

shock_fits_over_time: An array of fitted shock clusters over time.

ShockwaveDetection.calculate_normal_vector — Function

calculate_normal_vector(fit::Fitting, evenly_spaced_range, flow_data, t)

Calculate the normal vector to a fitted line model at a specific time step in the flow data.

Arguments

fit::Fitting: The fitting object containing the model and parameters of the fitted line.

evenly_spaced_range: A range of evenly spaced points along the x-axis.

flow_data: The flow data object containing the density field, number of cells, and bounds.

t: The specific time step at which to calculate the normal vector.

Returns

normal_quantity_x: The x-component of the normal vector.

normal_quantity_y: The y-component of the normal vector.

index_of_mid_x: The index of the midpoint in the x-direction.

index_of_mid_y: The index of the midpoint in the y-direction.

coordinate_movement_x: The movement in the x-direction based on the slope.

coordinate_movement_y: The movement in the y-direction based on the slope.

Description

This function calculates the normal vector to a fitted line model at a specific time step in the flow data. It first extracts the density field at the given time step and retrieves the number of cells and bounds from the flow data. If the model is a line model, it calculates the slope (m) and intercept (b) from the fit parameters. The magnitude of the normal vector is computed, and the x and y components of the normal vector are determined. The function then finds the midpoint of the line and its corresponding indices in the x and y directions. Finally, it calculates the coordinate movements based on the slope. This function is called within the create_heatmap_evo_with_shock function to calculate the normal vectors for the fitted shock models.

Visualization functions

ShockwaveDetection.create_heatmap_evo_with_shock — Function

create_heatmap_evo_with_shock(flow_data, detection, field::Symbol = :density_field, show_curve = true, show_normal_vector = true; T=Float64)

Create a heatmap evolution plot with shock detection for 1D or 2D flow data.

Arguments

flow_data: A FlowData object containing the flow field data.

detection: A detection object containing shock detection results.

field::Symbol: The field to visualize. Default is :density_field.

show_curve::Bool: Whether to show the shock curve in the plot. Default is true.

show_normal_vector::Bool: Whether to show the normal vector in the plot. Default is true.

T: The data type of the flow field. Default is Float64.

Description

This function creates a heatmap evolution plot with shock detection for either 1D or 2D flow data. It determines the dimensionality of the flow data and calls the appropriate helper function to generate the plot.

For 1D flow data (Array{T, 3}), it calls create_heatmap_evo_with_shock_1D.

For 2D flow data (Array{T, 4}) or from flow data of .celltape files with obstacles, it calls create_heatmap_evo_with_shock_2D.

ShockwaveDetection.create_wave_animation — Function

create_wave_animation(flow_data::FlowData)

Create an animation of density, velocity, and pressure fields over time from properties of the flow.

Arguments

flow_data::FlowData: A struct containing the flow data of x values, time steps, density field, velocity field, and pressure field.

Returns

This function saves an animation as a gif file named "densityvelocitypressureovertime.gif" in the current directory and returns the animated object.

ShockwaveDetection.create_wave_animation_with_shock — Function

create_wave_animation_with_shock(flow_data::FlowData, shock_positions_over_time)

Create an animation of density, velocity, and pressure fields over time from properties of the flow with shock wave positions.

Arguments

flow_data::FlowData: A struct containing the flow data of x values, time steps, density field, velocity field, and pressure field.

shock_positions_over_time::Vector{Vector{Int}}: A vector of vectors containing the shock positions over time from the detection algorithm.

Returns

This function saves an animation as a gif file named "densityvelocitypressureovertimewithshock_positions.gif" in the current directory

and returns the animated object.

ShockwaveDetection.plot_shock_fits_over_time — Function

plot_shock_fits_over_time(flow_data, detection, show_normal_vector = false; T=Float64)

Plot the shock fits over time for a given flow data and detection results. This differs from other visualization function in that it shows the shock fits and doesn't save it as gif file.

Arguments

flow_data: The flow data containing information about the flow field.

detection: An object containing the shock detection results, including shock clusters and fits over time.

show_normal_vector: A boolean flag indicating whether to show the normal vector for each shock fit. Defaults to false.

T: The type of the elements in the flow data's velocity field. Defaults to Float64.